Separate cells, headers and table into component files
