+++ frontend/src/features/people/components/index.ts
@@ -1,1 +1,1 @@
-export { default as Table } from './Table';
+export { default as Table } from './table';
