+++ frontend/src/features/people/pages/People.tsx
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import { Table } from '../components';
 
 const People = () => {
@@ -14,6 +15,19 @@ const People = () => {
     { accessor: 'age', header: 'Age' },
     // ... more columns
   ];
+
+  // Using state to dynamically manage columns
+  const [columns, setColumns] = useState(columnJson);
+
+  // Handler to add a new column
+  const handleAddColumn = () => {
+    // Example of a new column. You might want to make this more dynamic.
+    const newColumn = {
+      accessor: `newColumn${columns.length}`, // Ensure the accessor is unique
+      header: `New Column ${columns.length}`,
+    };
+    setColumns(prevColumns => [...prevColumns, newColumn]);
+  };
 
   return (
     <div className="w-full h-full container px-6 py-8">
